[Course statistics] Refactor utils
Move around util functions closer to the components they are only used in
