+++ services/frontend/src/constants/grades.ts
@@ -1,0 +1,6 @@
+export const FAILED_GRADES = ['eisa', 'hyl.', 'hyl', '0', 'luop']
+export const NUMERIC_GRADES = ['1', '2', '3', '4', '5']
+export const OTHER_PASSED_GRADES = ['hyv.', 'hyv']
+export const PASS_FAIL_GRADES = ['0', 'Hyv.']
+export const SECOND_NATIONAL_LANGUAGE_GRADES = ['TT', 'HT']
+export const THESIS_GRADES = ['I', 'A', 'NSLA', 'LUB', 'CL', 'MCLA', 'ECLA', 'L']
